WALL PLAN
Okay, I now need a way to check for wall collisions
I could store each wall as an array containing the coordinates for two points
That seems like a lot of work for each image of the map and I'm lazy
What if I instead, when moving an entity, I check if the pixels of the map image to the sides of the hitboxes of the entity are transparent?
If they are, then the object can move, if not then it can't?
I might also be able to basically shoot out a perpendicular ray from each of the corners of the hitbox to detemrine distance from the nearest wall.
I like this plan better, it seems easier
https://www.w3schools.com/jsref/canvas_getimagedata.asp